minor file renames
