--- packages/components/src/components/CodeEditor/languageSupport/squiggleLanguageExtension.ts
@@ -1,5 +1,0 @@
-import { squiggleLanguageSupport } from "./squiggle.js";
-
-export function squiggleLanguageExtension() {
-  return squiggleLanguageSupport();
-}
